Fix first-pass compile errors in the Windows engine core
- gsr_win32_compat.h: include <stdbool.h> in the force-included shim so
  translation units that only include the shim get bool (the compat
  implementation used it without the header).
- gsr_utils_win32.c: include capture/capture.h for the full
  gsr_capture_metadata definition and GSR_CAPTURE_ALIGN_* enums used by
  gsr_capture_get_target_position (utils.h only forward-declares the
  typedef). Mirrors upstream utils.c's include set.
- ci-smoke/main.c: DWORD is unsigned long on Windows (LLP64), so use
  %lu instead of %u for the OS version fields.

+++ tests/ci-smoke/main.c
@@ -36,8 +36,9 @@ int main(void) {
     if (ntdll) {
         RtlGetVersion_t fn = (RtlGetVersion_t)(void *)GetProcAddress(ntdll, "RtlGetVersion");
         if (fn && fn(&ovi) == 0)
-            printf("windows: %u.%u.%u (build %u)\n", ovi.dwMajorVersion, ovi.dwMinorVersion,
-                   ovi.dwBuildNumber, ovi.dwBuildNumber);
+            /* DWORD is unsigned long on Windows (LLP64), so %lu is correct. */
+            printf("windows: %lu.%lu.%lu (build %lu)\n", (unsigned long)ovi.dwMajorVersion, (unsigned long)ovi.dwMinorVersion,
+                   (unsigned long)ovi.dwBuildNumber, (unsigned long)ovi.dwBuildNumber);
     }
 #endif
